+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file. See [standa
   * Visual improvements to log creation page
   * Run types are now added and can be filtered. The type for a run is shown in the details page.
   * Filling filters for log overview will not automatically remove whitespaces on typing
+  * Time between runs is displayed on LHC fill display
 * Notable changes for developers:
   * Balloon system is now a global popover system, not limited to content overflow
   * Runs API: